test: split SafariCookieQueryStrategy extended tests into separate files
