+++ src/core/browsers/safari/__tests__/SafariCookieQueryStrategy.extended.test.ts
@@ -1,119 +1,5 @@
-import { homedir } from "os";
-
-import type { BinaryCookieRow } from "../../../../types/schemas";
-import { decodeBinaryCookies } from "../decodeBinaryCookies";
-import { SafariCookieQueryStrategy } from "../SafariCookieQueryStrategy";
-
-// Mock decodeBinaryCookies
-jest.mock("../decodeBinaryCookies");
-
-// Mock os.homedir and path.join
-jest.mock("os", () => ({
-  homedir: jest.fn(),
-}));
-
-jest.mock("path", () => ({
-  join: jest.fn((...args) => args.join("/")),
-}));
-
-describe("SafariCookieQueryStrategy - Extended Properties", () => {
-  let strategy: SafariCookieQueryStrategy;
-  const mockDecodeBinaryCookies = decodeBinaryCookies as jest.MockedFunction<
-    typeof decodeBinaryCookies
-  >;
-  const mockHomedir = homedir as jest.MockedFunction<typeof homedir>;
-
-  beforeEach(() => {
-    strategy = new SafariCookieQueryStrategy();
-    mockHomedir.mockReturnValue("/Users/testuser");
-    jest.clearAllMocks();
-  });
-
-  it("should handle cookies with comments", async () => {
-    const mockCookies: BinaryCookieRow[] = [
-      {
-        name: "test-cookie",
-        value: "test-value",
-        domain: "example.com",
-        path: "/",
-        creation: 1234567890,
-        expiry: 1234567890,
-        comment: "Test comment",
-        commentURL: "https://example.com/cookie-info",
-      },
-    ];
-
-    mockDecodeBinaryCookies.mockReturnValue(mockCookies);
-
-    const cookies = await strategy.queryCookies("test-cookie", "example.com");
-    expect(cookies).toHaveLength(1);
-    expect(cookies[0].meta?.comment).toBe("Test comment");
-    expect(cookies[0].meta?.commentURL).toBe("https://example.com/cookie-info");
-  });
-
-  it("should handle cookies with port numbers", async () => {
-    const mockCookies: BinaryCookieRow[] = [
-      {
-        name: "test-cookie",
-        value: "test-value",
-        domain: "example.com",
-        path: "/",
-        creation: 1234567890,
-        expiry: 1234567890,
-        port: 8080,
-      },
-    ];
-
-    mockDecodeBinaryCookies.mockReturnValue(mockCookies);
-
-    const cookies = await strategy.queryCookies("test-cookie", "example.com");
-    expect(cookies).toHaveLength(1);
-    expect(cookies[0].meta?.port).toBe(8080);
-  });
-
-  it("should handle cookies with version", async () => {
-    const mockCookies: BinaryCookieRow[] = [
-      {
-        name: "test-cookie",
-        value: "test-value",
-        domain: "example.com",
-        path: "/",
-        creation: 1234567890,
-        expiry: 1234567890,
-        version: 1,
-      },
-    ];
-
-    mockDecodeBinaryCookies.mockReturnValue(mockCookies);
-
-    const cookies = await strategy.queryCookies("test-cookie", "example.com");
-    expect(cookies).toHaveLength(1);
-    expect(cookies[0].meta?.version).toBe(1);
-  });
-
-  it("should handle cookies with undefined extended properties", async () => {
-    const mockCookies: BinaryCookieRow[] = [
-      {
-        name: "test-cookie",
-        value: "test-value",
-        domain: "example.com",
-        path: "/",
-        creation: 1234567890,
-        expiry: 1234567890,
-        comment: undefined,
-        commentURL: undefined,
-        port: undefined,
-        version: undefined,
-      },
-    ];
-
-    mockDecodeBinaryCookies.mockReturnValue(mockCookies);
-
-    const cookies = await strategy.queryCookies("test-cookie", "example.com");
-    expect(cookies).toHaveLength(1);
-    expect(cookies[0].meta?.comment).toBeUndefined();
-    expect(cookies[0].meta?.commentURL).toBeUndefined();
-    expect(cookies[0].meta?.port).toBeUndefined();
-    expect(cookies[0].meta?.version).toBeUndefined();
-  });
-});
+// This file has been split into separate test files:
+// - SafariCookieQueryStrategy.comments.test.ts
+// - SafariCookieQueryStrategy.port.test.ts
+// - SafariCookieQueryStrategy.version.test.ts
+// - SafariCookieQueryStrategy.undefined.test.ts
