+++ src/core/browsers/safari/__tests__/SafariCookieQueryStrategy.port.test.ts
@@ -1,0 +1,51 @@
+import { homedir } from "os";
+
+import type { BinaryCookieRow } from "../../../../types/schemas";
+import { decodeBinaryCookies } from "../decodeBinaryCookies";
+import { SafariCookieQueryStrategy } from "../SafariCookieQueryStrategy";
+
+// Mock decodeBinaryCookies
+jest.mock("../decodeBinaryCookies");
+
+// Mock os.homedir and path.join
+jest.mock("os", () => ({
+  homedir: jest.fn(),
+}));
+
+jest.mock("path", () => ({
+  join: jest.fn((...args) => args.join("/")),
+}));
+
+describe("SafariCookieQueryStrategy - Port Properties", () => {
+  let strategy: SafariCookieQueryStrategy;
+  const mockDecodeBinaryCookies = decodeBinaryCookies as jest.MockedFunction<
+    typeof decodeBinaryCookies
+  >;
+  const mockHomedir = homedir as jest.MockedFunction<typeof homedir>;
+
+  beforeEach(() => {
+    strategy = new SafariCookieQueryStrategy();
+    mockHomedir.mockReturnValue("/Users/testuser");
+    jest.clearAllMocks();
+  });
+
+  it("should handle cookies with port numbers", async () => {
+    const mockCookies: BinaryCookieRow[] = [
+      {
+        name: "test-cookie",
+        value: "test-value",
+        domain: "example.com",
+        path: "/",
+        creation: 1234567890,
+        expiry: 1234567890,
+        port: 8080,
+      },
+    ];
+
+    mockDecodeBinaryCookies.mockReturnValue(mockCookies);
+
+    const cookies = await strategy.queryCookies("test-cookie", "example.com");
+    expect(cookies).toHaveLength(1);
+    expect(cookies[0].meta?.port).toBe(8080);
+  });
+});
